+++ src/quick-lint-js/i18n/translation-table-generated.h
@@ -18,8 +18,8 @@ namespace quick_lint_js {
 using namespace std::literals::string_view_literals;
 
 constexpr std::uint32_t translation_table_locale_count = 5;
-constexpr std::uint16_t translation_table_mapping_table_size = 598;
-constexpr std::size_t translation_table_string_table_size = 82115;
+constexpr std::uint16_t translation_table_mapping_table_size = 599;
+constexpr std::size_t translation_table_string_table_size = 82149;
 constexpr std::size_t translation_table_locale_table_size = 35;
 
 QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
@@ -403,6 +403,7 @@ QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
           "missing 'break;' or '// fallthrough' comment between statement and 'case'"sv,
           "missing 'export' keyword for function"sv,
           "missing 'if' after 'else'"sv,
+          "missing 'new' in constructor type"sv,
           "missing 'while (condition)' for do-while statement"sv,
           "missing TypeScript type"sv,
           "missing arrow operator for arrow function"sv,